% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{bc4Squares}
\alias{bc4Squares}
\title{Plots a 4 squares plot}
\usage{
bc4Squares(
  bc,
  idents,
  lvl = NULL,
  top = 3,
  topnames = NULL,
  force = 1,
  alpha = 0.7,
  pt.size = 3
)
}
\arguments{
\item{bc}{\code{beyondcell} object.}

\item{idents}{Name of the metadata column of interest.}

\item{lvl}{Character vector with the \code{idents}' level of interest. If
\code{lvl = NULL}, all levels will be plotted.}

\item{top}{Number of top drugs per group to be labelled.}

\item{topnames}{Character vector with additional interesting drugs to be
labelled (either their names and/or sig IDs).}

\item{force}{(From \code{ggrepel}) Force of repulsion between overlapping
text labels. Defaults to 1.}

\item{alpha}{Transparency level between 1 (not transparent) and 0 (fully
transparent).}
}
\value{
A list of \code{ggplot2} objects, one per \code{lvl}.
}
\description{
This function drawns a 4 square plot of the drug signatures
present in a \code{\link[beyondcell]{beyondcell}} object. A 4 squares plot
consists in a scatter plot of the BCS' residuals (x axis) vs the switch
points (y axis). 4 quadrants are highlighted: the top-left and bottom-right
corners contain the drugs to which all selected cells are least/most
sensistive, respectively. The centre quadrants show the drugs to which half
of the selected cells are sensitive and the other half insensitive.

x cut-offs: first and last deciles; y cut-offs: 0.1, 0.4, 0.6 and 0.9.
}
\details{
This function returns a list of \code{\link[ggplot2]{ggplot2}}
objects, one per each \code{lvl}. Note that residuals are different for each
level while swicth points are signature-specific. So, x axis will vary and y
axis will remain constant accross all plots.
}
