% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{bcAddMetatada}
\alias{bcAddMetatada}
\title{Add new metadata columns to an existing beyondcell object}
\usage{
bcAddMetatada(bc, metadata = NULL)
}
\arguments{
\item{bc}{\code{beyondcell} object.}

\item{metadata}{Dataframe with metadata to add. Rownames should be cell names
and colnames should not be already present in \code{beyondcell@meta.data}.}
}
\value{
Returns a \code{beyondcell} object with new added metadata columns.
}
\description{
This function adds new metadata columns to an existing
\code{\link[beyondcell]{beyondcell}} object.
}
