% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{bcClusters}
\alias{bcClusters}
\title{Plots the UMAP reduction colored by metadata information}
\usage{
bcClusters(bc, UMAP = "beyondcell", idents, factor.col = TRUE, ...)
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{UMAP}{UMAP reduction to plot. Either "beyondcell" (computed using
\code{\link[bcUMAP]{bcUMAP}}) or "Seurat" computed using \code{Seurat}'s
functions.}

\item{idents}{Name of the metadata column to color by.}

\item{factor.col}{Logical indicating if \code{idents} column is a factor or
not. If \code{idents} is a numerical column (such as \code{percent.mt} or
\code{nFeature_RNA}, \code{factor.col} must be \code{FALSE}).}

\item{...}{Other arguments passed to \code{Seurat}'s
\code{\link[DimPlot]{DimPlot}}.}
}
\value{
A \code{\link[ggplot]{ggplot}} object with the UMAP reduction colored
by \code{idents}.
}
\description{
This function returns a {\link[ggplot]{ggplot}} object with the
UMAP reduction (either \code{beyondcell}'s or \code{Seurat}'s) colored by the
specified metadata column.
}
