% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{bcHistogram}
\alias{bcHistogram}
\title{Plots a histogram with the beyondcell scores of the signature of
interest}
\usage{
bcHistogram(bc, signatures = NULL, idents = NULL)
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{signatures}{Vector with the names of the signatures of interest. If
\code{signatures == "all"}, all signatures are selected.}

\item{idents}{Name of the metadata column of interest. If
\code{idents = NULL}, a single histogram with all bcscores is drawn. On the
other hand, if \code{idents != NULL} a histogram for each level found in
\code{idents} will be drawn.}
}
\value{
A list of \code{\link[ggplot]{ggplot}} histograms, one for each
signature of interest. In each histogram, the median, mean and sd are
reported. Also, the mean is indicated with a black dashed line and the median
with a red dashed line.
}
\description{
This function drawns a histogram plot of bcscores for each
signature of interest. The plot can be a single histogram (if
\code{idents = NULL}) or a histogram for each level found in \code{idents}.
}
