% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{bcMerge}
\alias{bcMerge}
\title{Merges two beyondcell objects}
\usage{
bcMerge(bc1, bc2)
}
\arguments{
\item{bc1}{First \code{beyondcell} object to merge. Can be a subset of 
\code{bc2}.}

\item{bc2}{Second \code{beyondcell} object to merge.}
}
\value{
A merged \code{beyondcell} object.
}
\description{
This function merges two \code{\link[beyondcell]{beyondcell}}
objects obtained from the same single-cell matrix using the same
\code{expr.thres} (see \code{\link[beyondcell]{bcScore}} for more
information). It binds signatures, not cells. If \code{bc1} is a subset of 
\code{bc2}, the resulting \code{beyondcell} object will contain just the 
subsetted cells and the \code{SeuratInfo} stored in \code{bc1}.
}
