% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{bcRecompute}
\alias{bcRecompute}
\title{Recomputes a beyondcell object}
\usage{
bcRecompute(bc, slot = "data")
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{slot}{Matrix to recompute the beyondcell object. Either \code{"data"}
or \code{"normalized"}.}
}
\value{
A recomputed \code{beyondcell} object.
}
\description{
This function recomputes a beyondcell object using the matrix
stored in the slot \code{@data} (original BCS) or \code{@normalized}
(which can contain subsetted and/or regressed BCS). Columns added with
\code{\link[beyondcell]{bcAddMetadata}} are preserved, except if they
define therapeutic clusters. Important: \code{bc@reductions} and
\code{bc@background} remain the same, while \code{bc@ranks} and
\code{bc@reductions} are removed.
}
