% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{bcScore}
\alias{bcScore}
\title{Computes the beyondcell score}
\usage{
bcScore(sc, gs, expr.thres = 0.1)
}
\arguments{
\item{sc}{\code{\link[Seurat]{Seurat}} object or expression matrix.}

\item{gs}{\code{\link[geneset]{geneset}} object.}

\item{expr.thres}{Minimum fraction of signature genes that must be
expressed in a cell to compute its beyondcell score. Cells with a number of
expressed genes below this fraction will have a \code{NaN} beyondcell
score.}
}
\value{
A \code{beyondcell} object.
}
\description{
This function computes the beyondcell score and returns an
object of class \code{\link[beyondcell]{beyondcell}}.
}
