% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{bcSignatures}
\alias{bcSignatures}
\title{Plots the UMAP reduction colored by bcscores or gene expression values}
\usage{
bcSignatures(
  bc,
  UMAP = "beyondcell",
  signatures = list(values = NULL, colorscale = NULL, alpha = 0.7, na.value = "grey50",
    limits = c(0, 1), center = NULL, breaks = 0.1),
  genes = list(values = NULL, limits = c(NA, NA), share.limits = FALSE),
  merged = NULL,
  blend = FALSE,
  mfrow = c(1, 1),
  ...
)
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{UMAP}{UMAP reduction to plot. Either \code{"beyondcell"} (computed
using \code{\link[bcUMAP]{bcUMAP}}) or \code{"Seurat"} computed using
\code{\link[Seurat]{Seurat}}'s functions.}

\item{signatures}{List with parameters to color the UMAP by bcscores:
\itemize{
\item{\code{values}:} {Vector with the names of the signatures of interest.
If \code{signatures[["values"]] = "all"}, all signatures are selected.}
\item{\code{colorscale}:} {Either a \code{viridis}, \code{RColorBrewer} or a
custom palette of 3 colors (low, medium and high) to color all signatures'
plots. If \code{colorscale = NULL} (default), the plots are colored using
\code{beyondcell}'s own palette.}
\item{\code{alpha}:} {Transparency level between 0 (not transparent) and 1
(fully transparent).}
\item{\code{na.value}:} {Color to use for missing values (\code{NA}s).}
\item{\code{limits}:} {Vector with the desired limits for all signatures'
plots.}
\item{\code{center}:} {A single number indicating the center of the
\code{colorscale} for all signatures' plots. If \code{center = NULL}
(default), the \code{center} for each signature is its switch point.}
\item{\code{breaks}:} {A single number indicating the break size of the
\code{colorscale}. Alternatively, it can be a vector with the desired breaks
(which don't have to be symmetric or equally distributed).}}}

\item{genes}{List with parameters to color the UMAP by gene expression
values:
\itemize{
\item{\code{values}:} {Vector with the names of the genes of interest. If
\code{genes[["values"]] = "all"}, all genes are selected.}
\item{\code{limits}:} {Vector with the desired limits for all genes' plots.
If \code{limits = c(NA, NA)} (default), the \code{limits} are computed for
each gene independently.}
\item{\code{share.limits}:} {Logical argument. If \code{share.limits = TRUE},
all genes' plots will have the same \code{limits}. If
\code{share.limits = FALSE} (default), each gene plot will have its own
\code{limits}. Note that if \code{limits != c(NA, NA)},
\code{share.limits = TRUE}.}}}

\item{merged}{If \code{merged != NULL}, two signatures will be superposed in
the same plot. If \code{merged = "direct"}, the signatures are assumed to
have a direct relationship and the bcscores will be added (+). On the other
hand, if \code{merged = "indirect"}, the signatures are assumed to have an
indirect relationship and their bcscores will be substracted (-).}

\item{blend}{(From \code{Seurat}) Scale and blend expression values to
visualize coexpression of two genes.}

\item{mfrow}{Numeric vector of the form \code{c(nr, nc)}. \code{nr}
corresponds to the number of rows and \code{nc} to the number of columns of
the arrays in which the plots will be drawn. If you want to draw the plots
individually, set \code{mfrow = c(1, 1)}.}

\item{...}{Other arguments passed to \code{Seurat}'s
\code{\link[FeaturePlot]{FeaturePlot}}.}
}
\value{
A list of \code{patchwork}s (if \code{mfrow != c(1, 1)}) or
\code{ggplot2}s (if \code{mfrow = c(1, 1)}) of the desired UMAP reduction
colored by the beyondcell scores (for signatures) or gene expression values
(for genes).
}
\description{
This function returns a list of
\code{\link[patchwork]{patchwork}}s or \code{\link[ggplot2]{ggplot2}}s with
the desired UMAP reduction (either \code{beyondcell}'s or \code{Seurat}'s)
colored by bcscores or gene expression values.
}
