% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{beyondcell-class}
\alias{beyondcell-class}
\alias{beyondcell}
\title{An S4 class to represent the beyondcell scores (BCS) for each cell and
signature.}
\description{
An S4 class to represent the beyondcell scores (BCS) for each cell and
signature.
}
\section{Slots}{

\describe{
\item{\code{scaled}}{(Subsetted and/or regressed) scaled BCS.}

\item{\code{normalized}}{(Subsetted and/or regressed) normalized BCS.}

\item{\code{data}}{Original normalized BCS, without subsetting or regression.}

\item{\code{switch.point}}{(Subsetted and/or regressed) scaled BCS for which the
normalized score in \code{@data} is 0 (one switch point per signature).}

\item{\code{ranks}}{List of dataframes with the statistics (switch point,
mean, median, sd, variance, min, max, proportion of \code{NaN}, residuals'
mean and ranks of each signature in the beyondcell object. This slot is
filled using the function \code{\link[beyondcell]{bcRanks}}.}

\item{\code{expr.matrix}}{Expression matrix used to compute the BCS.}

\item{\code{meta.data}}{Contains information about each cell (including the
therapeutic clusters and \code{Seurat}'s \code{@metadata}).}

\item{\code{background}}{(Subsetted and/or regressed) normalized BCS obtained using
DSS signatures. Useful to compute the UMAP reduction and the therapeutic
clusters when the number of drug signatures is low.}

\item{\code{reductions}}{A list of dimensional reductions for this object.}

\item{\code{regression}}{A list with the order of subset and regression steps
performed on the beyondcell object and the variables used for regression.}

\item{\code{n.genes}}{Argument passed to
\code{\link[beyondcell]{GenerateGenesets}}.}

\item{\code{mode}}{Argument passed to \code{GenerateGenesets}.}

\item{\code{thres}}{Argument \code{expr.thres} passed to
\code{\link[beyondcell]{bcScore}}.}
}}

