% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{colMinus}
\alias{colMinus}
\title{Substracts to the first row of a rectangular object the rest of rows}
\usage{
colMinus(x, na.rm = FALSE)
}
\arguments{
\item{x}{A matrix or a data frame.}

\item{na.rm}{(From \code{base}) Logical. Should missing values (including
\code{NaN}) be omitted from the calculations?}
}
\value{
A numeric rectangular object with the result of the substraction.
}
\description{
This function substracts to the first row of a numerical
rectangular object (\code{x[1, ]}) the rest of rows of the same rectangular
object (\code{x[2:length(x), ]}).
}
