% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{GeneCase}
\alias{GeneCase}
\title{Returns the fraction of each case type in the input vector}
\usage{
GeneCase(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
A named numeric vector with the fractions of each case type.
}
\description{
This function computes the fraction of of each case type
(uppercase, lowercase or capitalized) in a character vector.
}
