% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{GenerateGenesets}
\alias{GenerateGenesets}
\title{Creates a geneset object}
\usage{
GenerateGenesets(x, perform.reversal = FALSE)
}
\arguments{
\item{x}{A path to a GMT file. 
See Details for more information.}

\item{perform.reversal}{Logical. Return \code{beyoncell}'s pre-computed
signatures as if the signature was reversed.}
}
\value{
A \code{geneset} object.
}
\description{
This function creates a \code{\link[beyondcell]{geneset}}
object.
}
\details{
\code{x} can be:
\itemize{
\item{A path to a GMT file:} {A file that contains custom gene sets. Each
gene set must have an "_UP" or "_DOWN/_DN" suffix.}
}
}
