% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{GetCollection}
\alias{GetCollection}
\title{Load beyondcell collection}
\usage{
GetCollection(
  x,
  n.genes = 250,
  mode = c("up", "down"),
  filters = list(drugs = NULL, IDs = NULL, MoAs = NULL, targets = NULL, studies = NULL),
  include.pathways = TRUE
)
}
\arguments{
\item{x}{Either PSc, SSc or DSS.
See Details for more information.}

\item{n.genes}{Number of up and/or down-regulated genes used to compute each
signature.}

\item{mode}{Whether the output \code{geneset} must contain up and/or
down-regulated genes. See Details for more information.}

\item{filters}{If \code{x} is a pre-loaded collection, you can provide a list of
filters to subset it. You can specify which \code{drugs}, sig \code{IDs},
mechanisms of action (\code{MoAs}), \code{targets} and/or \code{studies} you
are interested in (cap insensitive). You can call
\code{\link[beyondcell]{ListFilters}} to check all the available values for
these filters. The signatures that pass \strong{ANY} of them are included in
the output.}

\item{include.pathways}{Logical. Return \code{beyoncell}'s pre-computed
signatures for functional pathways?}
}
\value{
A \code{geneset} object.
}
\description{
This function creates a \code{\link[beyondcell]{geneset}}
object from one of the default collections available.
}
\details{
\code{x} can be:
\itemize{
\item{A pre-loaded matrix:} {Either \code{PSc}, \code{SSc} or \code{DSS}.}
}
In addition, \code{mode} can be:
\itemize{
\item{\code{"up"}:} {To compute the signatures using only up-regulated
genes.}
\item{\code{"down"}:} {To compute the signatures using only down-regulated
genes.}
\item{\code{c("up", "down")} :} {To compute the signatures using both up and
down-regulated genes.}
}
}
