% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{ListFilters}
\alias{ListFilters}
\title{Returns all the possible values for the specified filter}
\usage{
ListFilters(entry)
}
\arguments{
\item{entry}{Either \code{"drugs"}, \code{"IDs"}, \code{"MoA"},
\code{"targets"} or \code{"source"}.}
}
\value{
All the possible values for the specified \code{entry}.
}
\description{
This function returns all the available values for \code{drugs},
\code{sig_ids}, \code{MoAs}, \code{targets} or \code{source} filters in
\code{\link[beyondcell]{GenerateGenesets}} function.
}
