% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{readGMT}
\alias{readGMT}
\title{Reads a GMT file as a list of gene sets}
\usage{
readGMT(x)
}
\arguments{
\item{x}{A path to a GMT file.}
}
\value{
A list of gene sets.
}
\description{
This function parses a GMT file and returns a list of gene sets.
}
