% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{bcClusters}
\alias{bcClusters}
\title{Plots a UMAP reduction coloured by metadata information}
\usage{
bcClusters(
  bc,
  idents,
  UMAP = "beyondcell",
  spatial = FALSE,
  factor.col = TRUE,
  mfrow = c(1, 1),
  ...
)
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{idents}{Name of the metadata column to colour by.}

\item{UMAP}{UMAP reduction to plot. Either \code{"beyondcell"}, computed
using \code{\link[beyondcell]{bcUMAP}}, or \code{"Seurat"}, obtained using
\code{Seurat}'s functions.}

\item{spatial}{Logical indicating if a \code{\link[Seurat]{SpatialPlot}}
should be drawn. If \code{spatial = TRUE}, the \code{UMAP} value will be 
ignored.}

\item{factor.col}{Logical indicating if \code{idents} column is a factor or
not. Set \code{factor.col = FALSE} if \code{idents} is a numeric column (such
as \code{percent.mt} or \code{nFeature_RNA}).}

\item{mfrow}{Numeric vector of the form \code{c(nr, nc)}. \code{nr}
corresponds to the number of rows and \code{nc} to the number of columns of
the grid in which the plots will be drawn. If you want to draw the plots
individually, set \code{mfrow = c(1, 1)}.}

\item{...}{Other arguments passed to 
\code{\link[Seurat]{DimPlot}}/\code{\link[Seurat]{SpatialDimPlot}} if 
\code{factor.col = TRUE} or 
\code{\link[Seurat]{FeaturePlot}}/\code{\link[Seurat]{SpatialFeaturePlot}} if 
\code{factor.col = FALSE}.}
}
\value{
A \code{ggplot2} object with the UMAP reduction coloured by
\code{idents}.
}
\description{
This function returns a {\link[ggplot2]{ggplot2}} object with
a UMAP reduction coloured by the specified metadata column.
}
