% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranks.R
\name{bcRanks}
\alias{bcRanks}
\title{Ranks the signatures from most sensitive to least sensitive}
\usage{
bcRanks(bc, idents = NULL, extended = TRUE)
}
\arguments{
\item{bc}{\code{\link[beyondcell]{beyondcell}} object.}

\item{idents}{Name of the metadata column of interest. The signatures' ranks 
are computed for each level in \code{idents}.}

\item{extended}{If \code{extended = TRUE}, this function returns the switch
point, mean, median, standard deviation, variance, min, max, proportion of
\code{NaN} and residuals' mean per signature. If \code{extended = FALSE},
this function returns only the switch point, mean and residuals' mean.}
}
\value{
A \code{beyondcell} object with the results in a new entry of
\code{bc@ranks}.
}
\description{
This function computes the beyondcell score's (BCS) statistics
of each signature and ranks them according to the switch point and 
residual's mean.
}
