% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.R
\name{center_scale_colour_stepsn}
\alias{center_scale_colour_stepsn}
\title{Creates a centered diverging binned colour gradient}
\usage{
center_scale_colour_stepsn(
  x,
  colorscale,
  alpha = 0.7,
  na.value = "grey50",
  limits = c(NA, NA),
  center = NULL,
  breaks = 0.1
)
}
\arguments{
\item{x}{A numeric vector. Can contain \code{NA}s.}

\item{colorscale}{A vector with 5 colors which can be obtained using
\code{\link[get_colour_stepsn]{get_colour_stepsn}}.}

\item{alpha}{Transparency level between 0 (not transparent) and 1 (fully
transparent).}

\item{na.value}{Color to use for missing values.}

\item{limits}{Vector with the desired limits.}

\item{center}{A single number indicating the center of the \code{colorscale}.
If \code{center = NULL} (default), the center is set to the middle point of
\code{x}.}

\item{breaks}{A single number indicating the break size of the
\code{colorscale}. Alternatively, it can be a vector with the desired breaks
(which don't have to be symmetric or equally distributed). If \code{center}
is a vector of two numbers, \code{breaks} are computed using the difference
between them.}
}
\value{
A centered diverging binned colour gradient that can be use to color
\code{\link[ggplot2]{ggplot2}} objects.
}
\description{
This function creates a diverging binned colour gradient
(low-mid-high) centered around \code{center}.
}
