% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{GenerateGenesets}
\alias{GenerateGenesets}
\title{Creates a geneset object}
\usage{
GenerateGenesets(
  x,
  n.genes = 250,
  mode = c("up", "down"),
  filters = list(drugs = NULL, IDs = NULL, MoA = NULL, targets = NULL, source = NULL),
  comparison = NULL,
  include.pathways = TRUE
)
}
\arguments{
\item{x}{A pre-loaded matrix, a numeric matrix or a path to a GMT file with
custom genesets.}

\item{n.genes}{Number of top and/or down genes in the desired genelists.}

\item{mode}{\code{"up"}, \code{"down"} or \code{c("up", "down")}. See Details
for more information.}

\item{filters}{If \code{x} is a pre-loaded matrix, you can provide a list of
filters to subset these matrices. You can specify which drug names, sig IDs,
mechanisms of action, target genes and sources you are interested in (cap
insensitive). You can call \code{\link[beyondcell]{ListFilters}} function to
check all the available values for these filters. The signatures that pass
\strong{ANY} of these filters are included in the output.}

\item{comparison}{\code{"treated_vs_control"} or
\code{"control_vs_treated"}. See Details for more information.}

\item{include.pathways}{\code{TRUE} (default) or \code{FALSE}. Whether or
not return \code{beyoncell}'s pre-computed genesets for functional pathways.}
}
\value{
A \code{\link[beyondcell]{geneset}} object with up and/or down genes for
each drug and pathway.
}
\description{
This function generates up and/or down genelists for each drug
and pathway.
}
\details{

}
