% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{GenerateGenesets}
\alias{GenerateGenesets}
\title{Creates a geneset object}
\usage{
GenerateGenesets(
  x,
  n.genes = 250,
  mode = c("up", "down"),
  filters = list(drugs = NULL, IDs = NULL, MoA = NULL, targets = NULL, source = NULL),
  comparison = NULL,
  include.pathways = TRUE
)
}
\arguments{
\item{x}{A pre-loaded matrix, a numeric matrix or a path to a GMT file with
custom genesets.}

\item{n.genes}{Number of top and/or down genes in the desired genelists.}

\item{mode}{\code{"up"}, \code{"down"} or \code{c("up", "down")}. See Details
for more information.}

\item{filters}{If \code{x} is a pre-loaded matrix, you can provide a list of
filters to subset these matrices. You can specify which drug names, sig IDs,
mechanisms of action, target genes and sources you are interested in (cap
insensitive). You can call \code{\link[ListFilters]{ListFilters}} function to
check all the available values for these filters. The signatures that pass
\strong{ANY} of these filters are included in the output.}

\item{comparison}{\code{"treated_vs_control"} or
\code{"sensitive_vs_resistant"}. See Details for more information.}

\item{include.pathways}{\code{TRUE} (default) or \code{FALSE}. Whether or
not return \code{beyoncell}'s pre-computed genesets for functional pathways.}
}
\value{
A \code{\link[geneset]{geneset}} object with up and/or down genes for
each drug and pathway.
}
\description{
This function generates up and/or down genelists for each drug
and pathway.
}
\details{
\code{x} can be:
\itemize{
\item{A pre-loaded matrix:} {Either \code{\link[PSc]{PSc}},
\code{\link[SSc]{SSc}} or \code{\link[DSS]{DSS}}.}
\item{A numeric matrix:} {A matrix with genes as rows and signatures as
columns that contains some type of numerical value such a t-stat or a LFC to
rank the genes accordingly.}
\item{A path to a GMT file:} {A file that contains custom genesets. Each
geneset must have an "_UP" or "_DOWN" suffix.}
}
In addition, \code{mode} can be:
\itemize{
\item{\code{"up"}:} {To return over-expressed genes in the signatures.}
\item{\code{"down"}:} {To return under-expressed genes in the signatures.}
}
If \code{x} is a path to a GMT file, \code{mode} is deprecated and the names
of all genesets must end in "_UP" or "_DOWN" to indicate the mode of each
one.

Finally, \code{comparison} can be:
\itemize{
\item{\code{"treated_vs_control"}:} {(\code{PSc} and \code{DSS} like) When
the numeric values or the genesets in the GMT file were obtained from a
comparison between drug treated and untreated cells.}
\item{\code{"sensitive_vs_resistant"}:} {(\code{SSc} like) When the numeric
values or the genesets in the GMT file were obtained from a comparison
between drug sensitive and resistant cells.}
}
When \code{x} is a pre-loaded matrix, \code{comparison} is set automatically.
}
