% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{geneset-class}
\alias{geneset-class}
\alias{geneset}
\title{An S4 class to represent the genesets of each drug}
\description{
An S4 class to represent the genesets of each drug
}
\section{Slots}{

\describe{
\item{\code{genelist}}{A list of drug and functional signatures (\code{pathways})
with up and/or down-regulated genes.}

\item{\code{n.genes}}{Argument passed to \code{\link[beyondcell]{GenerateGenesets}}.}

\item{\code{mode}}{Argument passed to \code{GenerateGenesets}.}

\item{\code{info}}{If \code{GenerateGenesets}' input is a pre-loaded matrix,
\code{info} will be a \code{data.frame} with the correspondences between
\code{sig_ids} and drug names, as well as the source of the pre-loaded
matrices (LINCS, CTRP, GDSC or CCLE). If \code{GenerateGenesets}' input is a
path to a GMT file, \code{info} will be empty.}

\item{\code{comparison}}{Argument passed to \code{GenerateGenesets}.}
}}

