% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genesets.R
\name{GetIDs}
\alias{GetIDs}
\title{Returns the \code{sig_ids} that match the specified values}
\usage{
GetIDs(values, filter, df = drugInfo)
}
\arguments{
\item{values}{User-supplied filtering vector for either drugs, MoAs, target
genes or source databases.}

\item{filter}{Column name to subset by. You can also spcify the colum}

\item{df}{\code{data.frame} with all drug information.}
}
\value{
A vector with the \code{sig_ids} that match the \code{filter}'s
elements.
}
\description{
This function subsets \code{df} to select only the entries that
match the specified \code{values} and returns the corresponding \code{sig_ids}.
}
